+++ akka-channels/src/main/scala/akka/channels/ChannelExtension.scala
